Added ES5 CommonJS functionality
