+++ test/test-classes/test-es6.js
@@ -1,0 +1,11 @@
+export class Test {
+  runTest() {
+    console.log('Running from test-es6.js!');
+  }
+}
+
+export default class Test2 {
+  runTest() {
+    console.log('Running from test-es6.js - default!');
+  }
+}
