--- test/test-classes/test.js
@@ -1,11 +1,0 @@
-export class Test {
-  runTest() {
-    console.log('Running from test.js!');
-  }
-}
-
-export default class Test2 {
-  runTest() {
-    console.log('Running from test.js - default!');
-  }
-}
